add methods and comment them

+++ src/Bank/CardAccount.java
@@ -6,18 +6,20 @@ public class CardAccount extends PaymentAccount {
         super();
     }
 
-    public CardAccount(long paymentScore) {
+    public CardAccount(double paymentScore) {
         super(paymentScore);
     }
 
     @Override
-    public void withdrawalCash(long amount) {
+    public void withdrawalCash(double amount) {
         if (requestBalance() < amount) {
             System.out.println("Current balance " + requestBalance() + " not enough to withdraw " + amount);
         } else {
-            long bankPercent = amount / 100;
+            double bankPercent = amount / 100;
+//            double bankPercentSimple = 1.01; // (balance - amount - amount * 0.01) = balance - amount * 1.01
             if (amount <= 100000) {
                  super.changeBalance(requestBalance() - amount - bankPercent);
+//                 super.withdrawalCash((amount * bankPercentSimple));
                 System.out.println("Withdrawal from account " + amount + " Bank commission is " + bankPercent);
             } else System.out.println("Cash withdrawal limit exceeded. You cannot withdraw the amount " + amount);
         }
