+++ src/Bank/PaymentAccount.java
@@ -2,23 +2,25 @@ package Bank;
 
 public class PaymentAccount {
 
-    private long paymentScore;
+    private double paymentScore;
 
-    public PaymentAccount() {
+    public PaymentAccount() { // конструктор родительского класса с установкой 0 баланса при создании
         paymentScore = 0;
     }
 
-    public PaymentAccount(long paymentScore) { // конструктор родительского класса
+    public PaymentAccount(double paymentScore) { // конструктор родительского класса с установкой баланса при создании
         this.paymentScore = paymentScore;
     }
 
-    public void depositCash(long cashIn) { // метод для пополнения счета (внесение наличных)
+
+
+    public void depositCash(double cashIn) { // метод для пополнения счета (внесение наличных)
         paymentScore = paymentScore + cashIn;
         System.out.println("Entered " + cashIn + " Current balance " + requestBalance());
     }
 
 
-    public void withdrawalCash(long cashOut) { // метод для вывода средств (снятие наличных)
+    public void withdrawalCash(double cashOut) { // метод для вывода средств (снятие наличных)
         if (requestBalance() < cashOut) {
             System.out.println("Account balance " + requestBalance() + " not enough to withdraw " + cashOut);
         }
@@ -29,12 +31,12 @@ public class PaymentAccount {
         else System.out.println("Cash withdrawal limit exceeded. You cannot withdraw the amount " + cashOut);
     }
 
-    void changeBalance(long paymentScore) { // метод для обновления баланса в дочерних методах
+    void changeBalance(double paymentScore) { // метод для обновления баланса в дочерних методах
         this.paymentScore = paymentScore;
     }
 
 
-    public long requestBalance() { // метод для запроса баланса
+    public double requestBalance() { // метод для запроса баланса
         return paymentScore;
     }
